move package schedule to job
