+++ shardingsphere-scaling/shardingsphere-scaling-core/src/main/java/org/apache/shardingsphere/scaling/core/job/schedule/JobStatus.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.core.schedule;
+package org.apache.shardingsphere.scaling.core.job.schedule;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
